third_party/WebKit/LayoutTests/editing/execCommand/crash-line-break-after-outdent.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
// For http://wkb.ug/60778
test(() => assert_selection(
  [
    '<div contenteditable>',
      '<pre id="x">',
        '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
          '<pre id="x">',
            '<img><img>',
            '<ul>',
              '<li><b><img></b></li>',
              '<li><b>|<br></b></li>',
            '</ul>',
          '</pre>',
        '</blockquote>',
      '</pre>',
    '</div>',
  ],
  'outdent',
  [
    '<div contenteditable>',
      '<pre id="x">',
        '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
          '<pre id="x">',
            '<img><img>',
            '<ul>',
              '<li><b><img></b></li>',
            '</ul>',
            '<b>|<br></b>',
          '</pre>',
        '</blockquote>',
      '</pre>',
    '</div>',
  ]),
  'Crash in line break after outdent');
</script>
